requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
google-generativeai>=0.3.0